^r Extract TOML policy load/merge from resolve_credential_sources.go, move-only (gofmt, vet, build, package tests pass; declaration set verified identical; internal restructure, no behavior change)
Step 1 of the authresolve god-file split, fragmented to fit the PR shape
gate: resolve_policy_load.go takes the policy loading, merging, include
validation, and TOML-subset parsing declarations verbatim.
